stored procedure added in scripts\cosmosdb folder and renamed GetItemQueryable method
