+++ WebApp.Logger/Loggers/Providers/CosmosDbs/CosmosDbAuditLogRepository.cs
@@ -54,7 +54,7 @@ namespace WebApp.Logger.Loggers.Repositories
         public async Task RetentionAsync(DateTime dateTime)
         {
             string date = dateTime.ToString("yyyy-MM-dd");//'T'HH: mm:ss.SSS'Z'
-            await _auditRepository.GetItemQueryable(date, _logOption.Log.Audit.GetType().Name.ToString().ToLower());
+            await _auditRepository.DeleteAsync(date, _logOption.Log.Audit.GetType().Name.ToString().ToLower());
         }
     }
 }
