+++ WebApp.Logger/Loggers/Providers/CosmosDbs/CosmosDbRouteLogRepository.cs
@@ -53,7 +53,7 @@ namespace WebApp.Logger.Loggers.Repositories
         public async Task RetentionAsync(DateTime dateTime)
         {
             string date = dateTime.ToString("yyyy-MM-dd");//'T'HH: mm:ss.SSS'Z'
-            await _routeRepository.GetItemQueryable(date, _logOption.Log.Request.GetType().Name.ToString().ToLower());
+            await _routeRepository.DeleteAsync(date, _logOption.Log.Request.GetType().Name.ToString().ToLower());
         }
     }
 }
