+++ WebApp/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "WebAppConnection": "Data Source=BS-1002\\SQLEXPRESS;Initial Catalog=DotnetLoggerWrapper; User Id=sa; Password = 12345; Integrated Security=False;MultipleActiveResultSets=True"
+    "WebAppConnection": "Data Source=localhost;Initial Catalog=DotnetLoggerWrapper; User Id=sa; Password = 123456; Integrated Security=False;MultipleActiveResultSets=True"
   },
   "MongoDbSettings": {
     "ConnectionString": "mongodb://localhost:27017",
@@ -78,7 +78,7 @@
         "IgnoreSchemas": [],
         "IgnoreTables": [],
         "IgnoreColumns": [ "Id", "CreatedBy" ],
-        "MaskColumns": [ "Password" ,"motto"]
+        "MaskColumns": [ "Password", "motto" ]
       }
     },
     "Provider": {
