+++ spDeleteSqlLogItems.txt
@@ -1,0 +1,44 @@
+function spDeleteSqlLogItems(query) {
+     var collection = getContext().getCollection();
+     var collectionLink = collection.getSelfLink();
+     var response = getContext().getResponse();
+     var responseBody = {
+         deleted: 0,
+         continuation: true
+     };
+     if (!query) throw new Error("The query is undefined or null.");
+     tryQueryAndDelete();
+     function tryQueryAndDelete(continuation) {
+         var requestOptions = {continuation: continuation};
+         var isAccepted = collection.queryDocuments(collectionLink, query, {}, function (err, retrievedDocs, responseOptions) {
+             if (err) throw err;
+             if (retrievedDocs.length > 0) {
+                 console.log(2);
+                 tryDelete(retrievedDocs);
+             } else if (responseOptions.continuation) {
+                 tryQueryAndDelete(responseOptions.continuation);
+             } else {
+                 responseBody.continuation = false;
+                 response.setBody(responseBody);
+             }
+         });
+         if (!isAccepted) {
+             response.setBody(responseBody);
+         }
+     }
+     function tryDelete(documents) {
+         if (documents.length > 0) {
+             var isAccepted = collection.deleteDocument(documents[0]._self, {}, function (err, responseOptions) {
+                 if (err) throw err;
+                 responseBody.deleted++;
+                 documents.shift();
+                 tryDelete(documents);
+             });
+             if (!isAccepted) {
+                 response.setBody(responseBody);
+             }
+         } else {
+             tryQueryAndDelete();
+         }
+     }
+ }
